Separated __init__.py from main.py
